Added cookie session
Creates a new table, lychee_sessions, in which a sha1-hash
and a timestamp is stored. The hash is stored in a cookie,
SESSION, on the client and allows a user to stay signed in
for longer than the php session. The hash is a combination
of the expiry timestamp, the username hash, the passwordhash
and the identifier.