Added user management dialog, and removed specific buttons
There is now a "Manage Users" menu item, instead
of the specific Add and delete user items.
This menu lists all users as well as their role,
with a button to delete them and a button to add new
users. When pressing a user a list of all albums
is shown, with checkboxes for read and write(non-functional).